Avoid replace benchmarks
